Basic Android support:
pthread_setcancelstate and pthread_cancel do not exist on Android and are define patched to no-op. TODO: investigate why these are used as opposed to "clean" shutdowns and fix.

USB/Serial: Android does not allow processes to claim and open interfaces. UsbManager on Android will open it on behalf of the application (after the user grants permission), and then pass the file handle through IPC for usage.
Furthermore, Android does not support USB serial natively anyways. Thus is must be implemented in userspace using Android USB API (libusb wrapper), after access to the device has been granted.
This means the most reasonable solution is to pass an already open file handle through the open-zwave stack for SerialControllerImpl to use.